Move special command to the end of vimrc
Since I observed github doesn't display properly this file in the
normal view, I guess special chars in it are an obstacle to github
highlighting process.

Thus I moved this line to the end of the file in an attempt to get
the whole file visible in that normal view.