Close comments with " like string literals
Since last commit didn't fix the problem with github normal view
of the file, my current guess is that comments of .vimrc, which
starts by " might be considered as string literals and the
highlighting system might be truncating them to a maximum amount
of chars.

Thus, my current attempt on it consists of closing these comments
with " as if they were string literals.