Rename file
Since last commit was also unsuccesful to solve the problem with
github normal view for the file, my current guess is that for some
reason github only displays 10 lines of the file, no matter what
the content is.

Thus, in this attempt I give an extension for the file, just in
case github undervalues no-extension files.