Replace special chars for readability
Special chars which we see in vim as '^[' and '^M' represent
special keystrokes (<ESC> and <ENTER> respectively) that we can't
visualize properly in the README.markdown. Moreover they're
displayed in a weird manner.

Since README is only for reading purposes, I replaced those chars
with the '^[' and '^M' representations, which are read the same
way we see them in vim.